requirements.txt
pyyaml
py7zr
requests